+++ static/app/views/replays/detail/page.tsx
@@ -11,8 +11,6 @@ import HeaderPlaceholder from 'sentry/components/replays/header/headerPlaceholde
 import ReplayMetaData from 'sentry/components/replays/header/replayMetaData';
 import ShareButton from 'sentry/components/replays/shareButton';
 import SentryDocumentTitle from 'sentry/components/sentryDocumentTitle';
-import TimeSince from 'sentry/components/timeSince';
-import {IconCalendar} from 'sentry/icons';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import type {ReplayError, ReplayRecord} from 'sentry/views/replays/types';
@@ -61,23 +59,7 @@ function Page({children, orgSlug, replayRecord, projectSlug, replayErrors}: Prop
             ip_address: replayRecord.user.ip || '',
             id: replayRecord.user.id || '',
           }}
-          // this is the subheading for the avatar, so displayEmail in this case is a misnomer
-          displayEmail={
-            <div>
-              {replayRecord ? (
-                <TimeContainer>
-                  <IconCalendar color="gray300" size="xs" />
-                  <StyledTimeSince
-                    date={replayRecord.started_at}
-                    isTooltipHoverable
-                    unitStyle="regular"
-                  />
-                </TimeContainer>
-              ) : (
-                <HeaderPlaceholder width="80px" height="16px" />
-              )}
-            </div>
-          }
+          hideEmail
         />
       ) : (
         <HeaderPlaceholder width="100%" height="58px" />
@@ -114,16 +96,6 @@ const ButtonActionsWrapper = styled(Layout.HeaderActions)`
   @media (max-width: ${p => p.theme.breakpoints.medium}) {
     margin-bottom: 0;
   }
-`;
-
-const TimeContainer = styled('div')`
-  display: flex;
-  gap: ${space(0.5)};
-  align-items: center;
-`;
-
-const StyledTimeSince = styled(TimeSince)`
-  font-size: ${p => p.theme.fontSizeMedium};
 `;
 
 export default Page;
